src/main/java/hydrogengamma/model/modules/TextIdentity.java
package hydrogengamma.model.modules;

import hydrogengamma.model.parsers.standard.TilesContainer;
import hydrogengamma.model.parsers.standard.Variable;
import hydrogengamma.model.parsers.standard.computers.Module;
import hydrogengamma.model.variables.TextVariable;

public class TextIdentity implements Module<TextVariable> {
    @Override
    public TextVariable execute(TilesContainer container, Variable<?>... args) {
        return (TextVariable) args[0];
    }

    @Override
    public boolean verify(Variable<?>... args) {
        return args.length == 1 && args[0] instanceof TextVariable;
    }
}
